Refactor DefenseSystem into modular core, effects, and renderer components
